Tasks

+++ src/Controller/TaskController.php
@@ -18,17 +18,26 @@ class TaskController extends AbstractController
     #[Route("/tasks", name: "task_list")]
     public function listAction(TaskRepository $task)
     {
-        return $this->render('task/list.html.twig', ['tasks' => $task->findAll()]);
+        $this->denyAccessUnlessGranted('ROLE_USER');
+        $tasks = $task->findAll();
+        return $this->render
+        (
+            'task/list.html.twig',
+            ['tasks' => $tasks],
+        );
+
     }
 
 
     #[Route("/tasks/create", name: "task_create")]
     public function createAction(Request $request, EntityManagerInterface $em)
     {
+        $this->denyAccessUnlessGranted('create');
         $task = new Task();
         $form = $this->createForm(TaskForm::class, $task);
         $form->handleRequest($request);
         if ($form->isSubmitted() === true && $form->isValid() === true) {
+            $task->setUser($this->getUser());
             $em->persist($task);
             $em->flush();
             $this->addFlash('success', 'La tâche a été bien été ajoutée.');
@@ -42,15 +51,12 @@ class TaskController extends AbstractController
     #[Route("/tasks/{id}/edit", name: "task_edit")]
     public function editAction(Task $task, Request $request, EntityManagerInterface $em)
     {
+        $this->denyAccessUnlessGranted('edit', $task, "Vous ne pouvez pas éditer cette tâche");
         $form = $this->createForm(TaskForm::class, $task);
-
         $form->handleRequest($request);
-
         if ($form->isSubmitted() === true && $form->isValid() === true) {
             $em->flush();
-
             $this->addFlash('success', 'La tâche a bien été modifiée.');
-
             return $this->redirectToRoute('task_list');
         }
 
@@ -58,31 +64,37 @@ class TaskController extends AbstractController
             'form' => $form,
             'task' => $task,
         ]);
+
     }
 
 
     #[Route("/tasks/{id}/toggle", name: "task_toggle")]
     public function toggleTaskAction(Task $task, EntityManagerInterface $em)
     {
+        $this->denyAccessUnlessGranted('toggle', $task, "Seul le créateur de la tâche peut en changer son état");
         $task->toggle(!$task->isDone());
         $em->flush();
-
-        $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
-
+        $message = $task->isDone() === true ? "terminée" : "en cours";
+        $this->addFlash('success', sprintf('La tâche %s a bien été marquée %s.', $task->getTitle(), $message));
         return $this->redirectToRoute('task_list');
+
     }
 
-    /**
-     * @Route("/tasks/{id}/delete", name="task_delete")
-     */
-    #[Route("/tasks/{id}/delete", name: "task_delete")]
-    public function deleteTaskAction(Task $task, EntityManagerInterface $em) :RedirectResponse
+
+    public function deleteTaskAction(Task $task, EntityManagerInterface $em, Request $request)
+
     {
-        $em->remove($task);
-        $em->flush();
-
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
-
+        $this->denyAccessUnlessGranted('delete', $task, "Vous ne pouvez pas supprimer la tâche d'un autre utilisateur");
+        $submittedToken = $request->request->get('token');
+        if ($this->isCsrfTokenValid('delete-task', $submittedToken)) {
+            $em->remove($task);
+            $em->flush();
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+            return $this->redirectToRoute('task_list');
+        }
+        $this->addFlash('error', "Vous n'êtes pas autorisé à supprimer cette tâche");
         return $this->redirectToRoute('task_list');
+
     }
+
 }
